Moved files into the primary source folder, added license headers
