+++ app/src/main/res/raw/cookingwords.txt
